Refactor fetchers into classes
