+++ restaurants/fetchers/__init__.py
@@ -1,0 +1,11 @@
+from .base import BaseFetcher
+from .google_places import GooglePlacesFetcher
+from .gov_csv import GovCsvFetcher
+from .osm import OsmFetcher
+
+__all__ = [
+    "BaseFetcher",
+    "GooglePlacesFetcher",
+    "GovCsvFetcher",
+    "OsmFetcher",
+]
