+++ restaurants/fetchers/base.py
@@ -1,0 +1,6 @@
+class BaseFetcher:
+    """Base interface for restaurant data fetchers."""
+
+    def fetch(self, zip_codes: list[str], **opts):
+        """Fetch data for the given ZIP codes."""
+        raise NotImplementedError
