+++ tests/test_refresh_restaurants.py
@@ -4,10 +4,11 @@ import pytest
 os.environ.setdefault("GOOGLE_API_KEY", "DUMMY")
 
 from restaurants import refresh_restaurants as rr
+from restaurants.fetchers import google_places as gp
 
 
 def test_google_details_use_threadpool(monkeypatch):
-    monkeypatch.setattr(rr, "check_network", lambda: True)
+    monkeypatch.setattr(gp, "check_network", lambda: True)
 
     class DummyResp:
         def __init__(self, data):
@@ -52,7 +53,7 @@ def test_google_details_use_threadpool(monkeypatch):
         else:
             raise AssertionError("unexpected url " + url)
 
-    monkeypatch.setattr(rr.requests.sessions.Session, "get", dummy_get)
+    monkeypatch.setattr(gp.requests.sessions.Session, "get", dummy_get)
 
     executors = []
 
@@ -80,11 +81,11 @@ def test_google_details_use_threadpool(monkeypatch):
         for f in iterable:
             yield f
 
-    monkeypatch.setattr(rr, "ThreadPoolExecutor", DummyExecutor)
-    monkeypatch.setattr(rr, "as_completed", dummy_as_completed)
+    monkeypatch.setattr(gp, "ThreadPoolExecutor", DummyExecutor)
+    monkeypatch.setattr(gp, "as_completed", dummy_as_completed)
 
     rr.smb_restaurants_data.clear()
-    rr.fetch_google_places(["98501"])
+    gp.GooglePlacesFetcher().fetch(["98501"])
 
     assert executors
     assert len(executors[0].submitted) == 2
@@ -92,28 +93,28 @@ def test_google_details_use_threadpool(monkeypatch):
 
 
 def test_fetch_google_places_no_network(monkeypatch):
-    monkeypatch.setattr(rr, "check_network", lambda: False)
+    monkeypatch.setattr(gp, "check_network", lambda: False)
     with pytest.raises(SystemExit):
-        rr.fetch_google_places(["98501"])
+        gp.GooglePlacesFetcher().fetch(["98501"])
 
 
 def test_fetch_google_places_textsearch_error(monkeypatch):
-    monkeypatch.setattr(rr, "check_network", lambda: True)
+    monkeypatch.setattr(gp, "check_network", lambda: True)
 
     class DummySessionGet:
         def __call__(self, url, params=None, timeout=None):
             if "textsearch" in url:
-                raise rr.requests.RequestException("boom")
+                raise gp.requests.RequestException("boom")
             raise AssertionError("unexpected url " + url)
 
-    monkeypatch.setattr(rr.requests.sessions.Session, "get", DummySessionGet())
+    monkeypatch.setattr(gp.requests.sessions.Session, "get", DummySessionGet())
 
     with pytest.raises(SystemExit):
-        rr.fetch_google_places(["98501"])
+        gp.GooglePlacesFetcher().fetch(["98501"])
 
 
 def test_fetch_google_places_details_failure(monkeypatch):
-    monkeypatch.setattr(rr, "check_network", lambda: True)
+    monkeypatch.setattr(gp, "check_network", lambda: True)
 
     class DummyResp:
         def __init__(self, data):
@@ -141,8 +142,8 @@ def test_fetch_google_places_details_failure(monkeypatch):
             raise RuntimeError("boom")
         raise AssertionError("unexpected url " + url)
 
-    monkeypatch.setattr(rr.requests.sessions.Session, "get", dummy_get)
-    monkeypatch.setattr(rr.time, "sleep", lambda _x: None)
+    monkeypatch.setattr(gp.requests.sessions.Session, "get", dummy_get)
+    monkeypatch.setattr(gp.time, "sleep", lambda _x: None)
 
     class DummyFuture:
         def __init__(self, res):
@@ -160,16 +161,16 @@ def test_fetch_google_places_details_failure(monkeypatch):
         def __exit__(self, exc_type, exc, tb):
             pass
 
-    monkeypatch.setattr(rr, "ThreadPoolExecutor", DummyExecutor)
-    monkeypatch.setattr(rr, "as_completed", lambda it: it)
+    monkeypatch.setattr(gp, "ThreadPoolExecutor", DummyExecutor)
+    monkeypatch.setattr(gp, "as_completed", lambda it: it)
 
     with pytest.raises(SystemExit):
-        rr.fetch_google_places(["98501"])
+        gp.GooglePlacesFetcher().fetch(["98501"])
 
 
 def test_main_missing_api_key(monkeypatch):
     monkeypatch.setattr(rr, "GOOGLE_API_KEY", None)
-    monkeypatch.setattr(rr, "fetch_google_places", lambda *_a, **_k: (_ for _ in ()).throw(AssertionError))
+    monkeypatch.setattr(rr, "FETCHERS", [])
     with pytest.raises(SystemExit):
         rr.main(["--zips", "98501"])
